ps3: switch to a minimal python3 implementation of make_npdata
This adds a minimal implementation of the make_npdata tool in python3.
The only thing implemented is minimal requirements in order to create
a ISO.BIN.EDAT for PS3 packages containing a PSX image. No more no less.

This removes one of the dependencies on external projects as we no
longer depend on this C program that users would need to compile.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>