pop-fe: fetch PIC1/ICON0 from the original game directory
and not the imb_files directory as the latter will be the current
pop-fe directory if we have merged a multi-bin file.

Also rework how we handle images and avoid rescaling the images
or converting them into a bytestream until we are actually just
about to write them to a file or store them inside a PBP.
This avoids pre-mature downscaling of images causing poor
visuals on the PS3 XMB when viewing the game list.

Suggested-by: HiroTex
Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>