Allow specifying pic0/pic1 from the gamesdb
Instead of having a randomly selected image that sometimes is not
great, because randomly selected, allow to put curated links
to pic0 and pic1 in the game database.

If there is no pic specified in the database then we fall back to the
original "pick a random one".

It does lose some of the charm though not knowing what images would be
used. The feeling of winning the lottery when the randomly picked
image turned out to be great.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>