pop-fe: add an argument to have pop-fe write CD-DA tracks to EBOOT.PBP
These tracks are not used by the emulator running on PS3 since it instead
plays the ATRAC3 encoded audio blobs and not the CD-DA tracks.
So most of the time the presence of the CD-DA data would just make the
EBOOT.PBP file much bigger than it has to be.

Make it possible to override this and force pop-fe to still write
the full CDDA tracks to the EBOOT.PBP.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>