pop-fe-ps3: add an option to only write the data track to the eboot
and skip writing the cd-da tracks.  The music tracks will still play
becuase they are written as atrac4 audio, it is just that the raw cd-da
tracks will not be written to the eboot, making the eboot smaller at the cost
that you can no longer extract the full disk image from the eboot any more.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>