pop-fe: fix hanging game Fear Effect NTSC-U and others
When the TOC from cu2, I computed end of track as one frame too little
which caused some games, like Fear Effect NTSC-U, to hang at the PS1 logo.
This is now corrected and the games no longer hangs.

Thanks to kozarovv for debugging the issue and finding the root cause
of the hangs.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>