Add support to use CDDA tracks instead of ATRACK3 for PSP
This because some games (V-Rally 2) depends on the feature of the audio
player that IF you tell it to play audio track #n, then when that track
ends it should automatically start playing the next track.
This does not happen for ATRAC3 audio on PSP so we need to switch that game
to use the raw CDDA tracks instead.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>