Fix alignment of the first AT3 track in the EBOOT
We used to only add alignment padding AFTER each audio track, causing the
first track to be misaligned.
Change this to add alignment padding to 16 byte boundary before the first
to make it correctly aligned.
Also we do not need to add alignment between the tracks as these just contain
raw ATRAC3 frames and thus are guaranteed to be a multiple of 16 bytes
in lenght anyway.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>