Disable PSTITLEIMG for single disc game PSP conversions
Some single disc games only works on old versions on POPS for PSP
and those old versions on POPS do not support the PSTITLEIMG section.

So, when generating a EBOOT for PSP and if it is a single disc
then generate it witout a PSTITLEIMG and only a single PSISOIMG section
so that it is compatible with old POPS versions.

Mortal Kombat Trilogy is one such example which only works on
POPS v 3.02.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>