Allow different revisions to share the same device id.
Use a different key for the games database for different disks
that share the same game id.
For example SLUS00884 and SLUS00884_R1.1 to distinguish between the
original release and the r1.1 relese.

Where it matters, instead of using the key as the game id, use
games[key]['id] instead and have ['id'] all map to the 9 character
game id of the original r1.0 release.

This should ONLY affect games that are done like this in the db
(i.e. us version of star wars the phantom menace) for now and not
affect anything else.
It is possible/likely/certain that some things will break but I will
fix that as I encounter it.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>