Update layout.py and pic0-offset
Change layout/py to show what both PSP and PS3 will look like.

Update pic0-offset to be percentage based to indicate a positive or negative
shift.

Example:
  'pic0-offset': (0.3, 0.1)
means : shift the image 30% to the right in the indicated red box for pic0
and shift the image down by 10% in the indicated red box.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>