Try to extract a PNG from the file
Decryption works and the Hash matches
but the contentof what I thought would be a PNG is slightly wrong.

000000 40 d1 07 00 00 00 00 00 02 00 00 00 00 00 00 00  >@...............<
000010 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00  >................<
000020 ca 93 e9 0f f6 dd 56 29 f0 cc 07 00 00 04 00 00  >......V)........<
000030 4c b8 59 d4 0b 56 42 98 00 00 00 0d 49 48 44 52  >L.Y..VB.....IHDR<
000040 00 00 01 e0 00 00 01 e0 08 02 00 00 00 f2 b6 29  >...............)<

The first 0x38 bytes are wrong.
The IHDR header starts at 0x3c which means the PNG header
should start at offset 0x30   but it is not.

Do they have a special proprietary version of PNG?

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>