Can pick apart a DOCUMENT.DAT and put it back together again
But can not create or generate the files that should go into it.
The first 0x38 bytes of the "files" are some kind of header that the
emulator uses.
Trying to zero them out or modifying these bytes before putting
the DOCUMENT.DAT back together causes the PS3 to crash when opening
software manual  so some of these bytes are important.
We need to figure out which ones are important an what values they need
to have before we will be able to generate manuals.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>